+++ tf_rnn/layers/utils/__init__.py
@@ -1,6 +1,6 @@
 """@since 0.6.1
 """
 
-from .accumulator import Accumulator
+from .accumulator import Accumulator, AccumulatorData
 from .confusion_matrix import ConfusionMatrix
 from .metrics import Metrics